+++ models/db/__init__.py
@@ -1,4 +1,4 @@
-from client import DBClient
+from .client import DBClient
 
 Db = DBClient(host='localhost',port=5233,
               user='postgres',password='postgres',db='postgres')
