+++ models/menus/menu.py
@@ -5,7 +5,7 @@ from typing import Optional
 from sqlalchemy.orm import Mapped
 from sqlalchemy.orm import mapped_column
 from sqlalchemy.orm import relationship
-from ..object.base import Base
+from ..core.base import Base
 
 
 class Menu(Base):
