+++ models/redis/__init__.py
@@ -1,4 +1,4 @@
-from client import RedisClient
+from .client import RedisClient
 from app import app
 
 redis_client = RedisClient(app=app)
